requirements.txt
alembic==1.14.1
asn1crypto==1.5.1
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
pillow==11.1.0
pip3-autoremove==1.2.2
psycopg2-binary==2.9.10
pycparser==2.22
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==8.0
scramp==1.4.5
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
